+++ src/main/kotlin/com/github/anddd7/controller/ProductController.kt
@@ -20,12 +20,12 @@ class ProductController(private val productService: ProductService) {
   fun products(): Flux<Product> = productService.findAll()
 
   @GetMapping("/{id}")
-  fun productStock(@PathVariable id: Long): Mono<ProductStockDTO> =
+  fun productStock(@PathVariable id: Int): Mono<ProductStockDTO> =
       productService.getProductStock(id)
           .switchIfEmpty(Mono.error(NoSuchElementException()))
 
   @GetMapping("/{id}/stock")
-  fun stock(@PathVariable id: Long): Mono<BigDecimal> = Mono.just(
+  fun stock(@PathVariable id: Int): Mono<BigDecimal> = Mono.just(
       Random(id).nextFloat().let(::abs).toBigDecimal()
   )
 }
