+++ src/main/kotlin/com/github/anddd7/repository/ProductRepository.kt
@@ -8,10 +8,15 @@ import org.springframework.stereotype.Repository
 import reactor.core.publisher.Flux
 import reactor.core.publisher.Mono
 
+/**
+ * can replaced by ReactiveCrudRepository, like jpa repository
+ *
+ * @see org.springframework.data.repository.reactive.ReactiveCrudRepository
+ */
 @Repository
 class ProductRepository(private val databaseClient: DatabaseClient) {
   fun findAll(): Flux<Product> = databaseClient.select().from<Product>().fetch().all()
-  fun getOne(id: Long): Mono<Product> = databaseClient.select().from<Product>()
+  fun getOne(id: Int): Mono<Product> = databaseClient.select().from<Product>()
       .matching(
           where("id").`is`(id)
       )
