+++ src/main/resources/db/migration/V202001081658__dml_initial_product.sql
@@ -3,5 +3,5 @@ CREATE TABLE PRODUCTS
     id    SERIAL NOT NULL
         CONSTRAINT PRODUCTS_PK PRIMARY KEY,
     name  VARCHAR,
-    price REAL
+    price NUMERIC(19, 2)
 );
